refactoring and trial helper classes
